flow_pre_clean
Added scaladoc.

Moved code to package org.opennetworkinsight
