+++ src/main/scala/org/opennetworkinsight/FlowPreLDA.scala
@@ -1,16 +1,17 @@
 
-package main.scala
+package org.opennetworkinsight
 
-import main.scala.FlowTransformation
-import main.scala.{FlowColumnIndex => indexOf}
 import org.apache.log4j.{Level, Logger => apacheLogger}
-import org.apache.spark.{SparkConf, SparkContext}
 import org.apache.spark.rdd.RDD
-import org.slf4j.{LoggerFactory, Logger}
-
+import org.apache.spark.{SparkConf, SparkContext}
+import org.slf4j.LoggerFactory
 
 import scala.io.Source
 
+
+/**
+  * Contains routines for creating the "words" for a suspicious connects analysis from incoming netflow records.
+  */
 object FlowPreLDA {
 
   def run() = {
@@ -145,7 +146,7 @@ object FlowPreLDA {
 
     logger.info("Trying to read file:  " + file)
     val rawdata: RDD[String] = sc.textFile(file)
-    val datanoheader: RDD[String] = FlowTransformation.removeHeader(rawdata)
+    val datanoheader: RDD[String] = FlowWordCreation.removeHeader(rawdata)
 
     val scoredFileExists = new java.io.File(scoredFile).exists
 
@@ -165,7 +166,7 @@ object FlowPreLDA {
 
     val datagood: RDD[String] = totalData.filter(line => line.split(",").length == 27)
 
-    val data_with_time = datagood.map(_.trim.split(',')).map(FlowTransformation.addTime)
+    val data_with_time = datagood.map(_.trim.split(',')).map(FlowWordCreation.addTime)
 
     logger.info("calculating time cuts ...")
     time_cuts = Quantiles.distributedQuantilesQuant(Quantiles.computeEcdf(data_with_time.map(row => row(indexOf.NUMTIME).toDouble)))
@@ -177,9 +178,9 @@ object FlowPreLDA {
     ipkt_cuts = Quantiles.distributedQuantilesQuint(Quantiles.computeEcdf(data_with_time.map(row => row(16).toDouble)))
     logger.info(ipkt_cuts.mkString(","))
 
-    val binned_data = data_with_time.map(row => FlowTransformation.binIbytIpktTime(row, ibyt_cuts, ipkt_cuts, time_cuts))
+    val binned_data = data_with_time.map(row => FlowWordCreation.binIbytIpktTime(row, ibyt_cuts, ipkt_cuts, time_cuts))
 
-    val data_with_words = binned_data.map(row => FlowTransformation.adjustPort(row))
+    val data_with_words = binned_data.map(row => FlowWordCreation.adjustPort(row))
 
     //next groupby src to get src word counts
     val src_word_counts = data_with_words.map(row => (row(indexOf.SOURCEIP) + " " + row(indexOf.SOURCEWORD), 1)).reduceByKey(_ + _)
